+++ app/renderer/input.js
@@ -674,8 +674,9 @@ let inputHistoryIndex = 0
 let lastActionInMapping = null
 let lastExecutedMapstring = null
 let draggingScreenshotFrame = false
-let lastX = 0
-let lastY = 0
+let lastScreenshotX = 0
+let lastScreenshotY = 0
+let suggestionTimer = null
 
 const init = () => {
     window.addEventListener("keydown", handleKeyboard)
@@ -900,8 +901,8 @@ const resetScreenshotDrag = () => setTimeout(() => {
 }, 10)
 
 const moveScreenshotFrame = (x, y) => {
-    const deltaX = x - lastX
-    const deltaY = y - lastY
+    const deltaX = x - lastScreenshotX
+    const deltaY = y - lastScreenshotY
     if (getMouseConf("screenshotframe") && draggingScreenshotFrame) {
         const url = document.getElementById("url")
         const dims = url.value.split(" ").find(
@@ -929,10 +930,10 @@ const moveScreenshotFrame = (x, y) => {
         url.value = url.value.replace(/\d+,\d+,\d+,\d+/g, `${rect.width},${
             rect.height},${rect.x},${rect.y}`)
         updateScreenshotHighlight()
-        updateSuggestions()
+        requestSuggestUpdate()
     }
-    lastX = x
-    lastY = y
+    lastScreenshotX = x
+    lastScreenshotY = y
 }
 
 const cutInput = (event = null) => {
@@ -948,7 +949,7 @@ const cutInput = (event = null) => {
     url.value = url.value.substr(0, url.selectionStart)
         + url.value.substr(url.selectionEnd)
     url.setSelectionRange(cur, cur)
-    updateSuggestions()
+    requestSuggestUpdate()
     updateNavbarScrolling()
 }
 
@@ -971,7 +972,7 @@ const pasteInput = (event = null) => {
     url.value = url.value.substr(0, url.selectionStart) + pastedText
         + url.value.substr(url.selectionEnd)
     url.setSelectionRange(cur + pastedText.length, cur + pastedText.length)
-    updateSuggestions()
+    requestSuggestUpdate()
     updateNavbarScrolling()
 }
 
@@ -1749,7 +1750,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
             const histEntry = inputHistoryList[inputHistoryIndex]
             url.value = histEntry.value
             url.setSelectionRange(histEntry.index, histEntry.index)
-            updateSuggestions(false)
+            requestSuggestUpdate(false)
         }
         updateNavbarScrolling()
         return
@@ -1760,7 +1761,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
             const histEntry = inputHistoryList[inputHistoryIndex]
             url.value = histEntry.value
             url.setSelectionRange(histEntry.index, histEntry.index)
-            updateSuggestions(false)
+            requestSuggestUpdate(false)
         }
         updateNavbarScrolling()
         return
@@ -1779,7 +1780,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
         url.value = url.value.substr(0, url.selectionStart)
             + url.value.substr(url.selectionEnd)
         url.setSelectionRange(cur, cur)
-        updateSuggestions()
+        requestSuggestUpdate()
         updateNavbarScrolling()
         if (id === "<Del>" || id.endsWith("-Del>")) {
             return
@@ -1794,7 +1795,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
             url.value = `${url.value.substr(0, url.selectionStart)}${
                 url.value.substr(url.selectionEnd + 1)}`
             url.setSelectionRange(cur, cur)
-            updateSuggestions()
+            requestSuggestUpdate()
             updateNavbarScrolling()
         }
         return
@@ -1808,7 +1809,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
                 url.value = `${url.value.substr(0, url.selectionStart)}${
                     url.value.substr(wordPosition)}`
                 url.setSelectionRange(cur, cur)
-                updateSuggestions()
+                requestSuggestUpdate()
                 updateNavbarScrolling()
                 return
             }
@@ -1821,7 +1822,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
             url.value = `${url.value.substr(0, url.selectionStart - 1)}${
                 url.value.substr(url.selectionEnd)}`
             url.setSelectionRange(cur - 1, cur - 1)
-            updateSuggestions()
+            requestSuggestUpdate()
             updateNavbarScrolling()
         }
         return
@@ -1834,7 +1835,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
                 url.value = `${url.value.substr(0, wordPosition)}${
                     url.value.substr(url.selectionStart)}`
                 url.setSelectionRange(wordPosition, wordPosition)
-                updateSuggestions()
+                requestSuggestUpdate()
                 updateNavbarScrolling()
                 return
             }
@@ -1865,7 +1866,7 @@ const typeCharacterIntoNavbar = (character, force = false) => {
     }
     if (text !== url.value) {
         url.setSelectionRange(cur + 1, cur + 1)
-        updateSuggestions()
+        requestSuggestUpdate()
         updateNavbarScrolling()
     }
 }
@@ -1875,20 +1876,36 @@ const resetInputHistory = () => {
     inputHistoryIndex = 0
 }
 
-const updateSuggestions = (updateHistory = true) => {
+const requestSuggestUpdate = (updateHistory = true) => {
     const url = document.getElementById("url")
+    const suggestBounceDelay = getSetting("suggestbouncedelay")
     if (updateHistory) {
         inputHistoryList = inputHistoryList.slice(0, inputHistoryIndex + 1)
         inputHistoryIndex = inputHistoryList.length
         inputHistoryList.push({"index": url.selectionStart, "value": url.value})
     }
-    if (currentMode() === "explore") {
+    if (!suggestionTimer) {
+        updateSuggestions()
+    }
+    if (suggestBounceDelay) {
+        clearTimeout(suggestionTimer)
+        suggestionTimer = setTimeout(() => {
+            suggestionTimer = null
+            updateSuggestions()
+        }, suggestBounceDelay)
+    }
+}
+
+const updateSuggestions = () => {
+    const val = document.getElementById("url").value
+    const mode = currentMode()
+    if (mode === "explore") {
         const {suggestExplore} = require("./suggest")
-        suggestExplore(url.value)
-    } else if (currentMode() === "command") {
+        suggestExplore(val)
+    } else if (mode === "command") {
         const {suggestCommand} = require("./suggest")
-        suggestCommand(url.value)
-    } else if (currentMode() === "search" && getSetting("incsearch")) {
+        suggestCommand(val)
+    } else if (mode === "search" && getSetting("incsearch")) {
         ACTIONS.incrementalSearch()
     }
 }
@@ -2206,12 +2223,12 @@ module.exports = {
     moveScreenshotFrame,
     pasteInput,
     repeatLastAction,
+    requestSuggestUpdate,
     resetInputHistory,
     resetScreenshotDrag,
     sanitiseMapString,
     typeCharacterIntoNavbar,
     uncountableActions,
     unmap,
-    updateKeysOnScreen,
-    updateSuggestions
+    updateKeysOnScreen
 }